ingestion/bq_event_loader/requirements.txt
flask==3.0.2
google-cloud-bigquery==3.27.0
gunicorn==22.0.0